requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.3.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
distro==1.9.0
Django==4.2.27
django-allauth==65.13.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
firebase-admin==6.5.0
fonttools==4.53.1
frozenlist==1.6.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-vision
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
importlib_resources==6.4.0
jdcal==1.4.1
jiter==0.9.0
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.9.1
msgpack==1.0.8
multidict==6.4.3
numpy==2.0.0
openai==0.28.0
openpyexcel==2.5.14
packaging==24.1
pillow==10.4.0
propcache==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.8.0
pyparsing==3.1.2
pyTelegramBotAPI==4.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2023.3
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SpeechRecognition==3.14.3
sqlparse==0.4.4
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.39.0
yarl==1.20.0
zipp==3.19.2
